IGNITE-15629 Code cleanup
